Add FSM for food retrieval
Add subscriber to listen for requests
Add FSM that directs the robot to go to vendors and home
Inflate map boundaries
Add timeout to AStar. Sometimes AStar breaks out of the map and runs forever if location around goal is not free
Add error handling when planned location is not free. Still need to get this to work